EntityLogs : ajout interface + gestion json directe
