+++ sdk_entrepot_gpf/store/ProcessingExecution.py
@@ -1,13 +1,13 @@
 from datetime import datetime
-import json
 from typing import Optional
 
 from sdk_entrepot_gpf.store.StoreEntity import StoreEntity
 from sdk_entrepot_gpf.store.interface.CsfInterface import CsfInterface
+from sdk_entrepot_gpf.store.interface.LogsInterface import LogsInterface
 from sdk_entrepot_gpf.io.ApiRequester import ApiRequester
 
 
-class ProcessingExecution(CsfInterface, StoreEntity):
+class ProcessingExecution(CsfInterface, LogsInterface, StoreEntity):
     """Classe Python représentant l'entité ProcessingExecution (exécution d'un traitement).
 
     Cette classe permet d'effectuer les actions spécifiques liées aux exécution de traitement : création,
@@ -23,31 +23,6 @@ class ProcessingExecution(CsfInterface, StoreEntity):
     STATUS_SUCCESS = "SUCCESS"
     STATUS_FAILURE = "FAILURE"
     STATUS_ABORTED = "ABORTED"
-
-    def api_logs(self) -> str:
-        """Récupère les logs de cette exécution de traitement sur l'API.
-
-        Returns:
-            str: les logs récupérés
-        """
-        # Génération du nom de la route
-        s_route = f"{self._entity_name}_logs"
-        # Requête "get"
-        o_response = ApiRequester().route_request(
-            s_route,
-            route_params={self._entity_name: self.id, "datastore": self.datastore},
-        )
-        s_log = o_response.text
-        try:
-            if s_log in ["", "[]"]:
-                return ""
-            # les logs sont retourné sous forme de liste on tente le passage de liste à un texte propre.
-            l_log = json.loads(s_log)
-            s_log = "\n".join(l_log)
-        except Exception:
-            pass
-        # on renvoie les logs
-        return s_log
 
     def api_launch(self) -> None:
         """Lance l'exécution du traitement sur l'API."""
